Improve production docker build - make better use of layer cache
There is no need to bust the "npm install" cache layer if only the app
source files have changed. This only applies to the prod image.
